warnings fix

+++ CPP/tabu_search.h
@@ -11,5 +11,5 @@ struct Conflict
     int node[2];
     int colour;
 };
-std::vector<Conflict> how_many_conflicts(const MatrixGraph &G,const std::vector<std::vector<int>> &solution);
-std::vector<std::vector<int>> tabu_search(MatrixGraph G, int k_number_of_colours, std::vector<std::vector<int>> solution, int tabu_size = 7, int number_of_neighbours = 10, int max_iterations = 100000);
+std::vector<Conflict> how_many_conflicts(const MatrixGraph &G, const std::vector<std::vector<int>> &solution);
+std::vector<std::vector<int>> tabu_search(MatrixGraph G, int k_number_of_colours, std::vector<std::vector<int>> solution, size_t tabu_size = 7, int number_of_neighbours = 10, int max_iterations = 100000);
